fix(event-streams): close review follow-ups
