+++ crates/ironclaw_event_streams/Cargo.toml
@@ -12,6 +12,7 @@ ironclaw_event_projections = { path = "../ironclaw_event_projections" }
 ironclaw_host_api = { path = "../ironclaw_host_api" }
 ironclaw_outbound = { path = "../ironclaw_outbound" }
 ironclaw_turns = { path = "../ironclaw_turns" }
+parking_lot = "0.12"
 serde = { version = "1", features = ["derive", "rc"] }
 serde_json = "1"
 thiserror = "2"
