+++ docs/assumptions/reborn-event-stream-manager.md
@@ -14,4 +14,4 @@ The first slice keeps production transports and DB-backed projection stream stor
 Review hardening added host-owned limits because the issue does not define final production sizing:
 
 - Product subscribers may request a buffer capacity, but the host caps it at 128 items and clamps zero to one item.
-- Redaction-validation decisions are cached per envelope kind, scope, and runtime cursor. The in-memory cache is capped at 1024 decisions and clears on overflow; a later durable/multitenant stream host can replace that policy with a tenant-aware LRU if needed.
+- Redaction-validation decisions are cached per envelope kind, scope, runtime cursor, payload length, and payload digest. The in-memory cache is capped at 1024 decisions and evicts incrementally by recency on overflow; a later durable/multitenant stream host can replace that policy with a tenant-aware LRU if needed.
